# openapi.yml
openapi: 3.1.0
info:
  title: Masashi Enokida API
  version: 1.0.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        role:
          type: string
          enum: [USER, MEMBER_FREE, MEMBER_GOLD, ADMIN]
    Subscription:
      type: object
      required: [hasSubscription, tier]
      properties:
        hasSubscription:
          type: boolean
        tier:
          type: string
          enum: [USER, MEMBER_FREE, MEMBER_GOLD]
        status:
          type: string
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
        cancelAtPeriodEnd:
          type: boolean
    AuthMeResponse:
      type: object
      required: [user, subscription]
      properties:
        user:
          $ref: '#/components/schemas/User'
        subscription:
          $ref: '#/components/schemas/Subscription'
...
    BiographyEntry:
      type: object
      required: [id, year, description, sort_order]
      properties:
        id:
          type: string
          format: uuid
        year:
          type: string
        description:
          type: string
        sort_order:
          type: integer
    BiographyInput:
      type: object
      required: [year, description]
      properties:
        year:
          type: string
        description:
          type: string
        sort_order:
          type: integer
    BlogPostSummary:
      type: object
      required: [id, title, publishedAt, membersOnly]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        excerpt:
          type: string
          nullable: true
        thumbnail:
          type: object
          nullable: true
          properties:
            url:
              type: string
        category:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
        publishedAt:
          type: string
          format: date-time
        membersOnly:
          type: boolean
        isLocked:
          type: boolean
    BlogPost:
      allOf:
        - $ref: '#/components/schemas/BlogPostSummary'
        - type: object
          properties:
            content:
              type: string
              nullable: true
    BlogPostList:
      type: object
      required: [posts, totalPages, total]
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/BlogPostSummary'
        totalPages:
          type: integer
        total:
          type: integer
...
    AdminBiographyEntry:
      type: object
      required: [id, year, description, sort_order, created_at]
      properties:
        id:
          type: string
          format: uuid
        year:
          type: string
        description:
          type: string
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
...
    AdminBlogCategory:
      type: object
      required: [id, name, slug, sort_order, post_count, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        sort_order:
          type: integer
        post_count:
          type: integer
        created_at:
          type: string
          format: date-time
    AdminBlogCategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        sort_order:
          type: integer
    AdminBlogPostSummary:
      type: object
      required: [id, title, is_published, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        excerpt:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        category_id:
          type: string
          format: uuid
          nullable: true
        members_only:
          type: boolean
        is_published:
          type: boolean
        published_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    AdminBlogPost:
      type: object
      required: [id, title, content, is_published, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        category_id:
          type: string
          format: uuid
          nullable: true
        members_only:
          type: boolean
        is_published:
          type: boolean
        published_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    AdminBlogPostInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        category_id:
          type: string
          format: uuid
          nullable: true
        members_only:
          type: boolean
          default: false
        is_published:
          type: boolean
          default: false
        published_at:
          type: string
          format: date-time
          nullable: true
    AdminBlogPostList:
      type: object
      required: [posts, total, totalPages]
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/AdminBlogPostSummary'
        total:
          type: integer
        totalPages:
          type: integer
paths:
...
  /auth/me:
    get:
      operationId: getAuthMe
      summary: Get current user and subscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          description: Unauthorized
  /auth/signout:
    post:
      operationId: postAuthSignout
      summary: Sign out
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /auth/account:
    delete:
      operationId: deleteAuthAccount
      summary: Delete account
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
  /biography:
    get:
      operationId: getBiography
      summary: List biography entries
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BiographyEntry'
    post:
      operationId: createBiography
      summary: Create biography entry (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiographyInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiographyEntry'
  /biography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateBiography
      summary: Update biography entry (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiographyInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiographyEntry'
    delete:
      operationId: deleteBiography
      summary: Delete biography entry (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /biography/reorder:
    put:
      operationId: reorderBiography
      summary: Reorder biography entries (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [id, sort_order]
                properties:
                  id:
                    type: string
                    format: uuid
                  sort_order:
                    type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
  /blog:
    get:
      operationId: getBlog
      summary: List blog posts (with member gating)
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPostList'
  /blog/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getBlogPost
      summary: Get single blog post (with member gating)
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          description: Not found
...
  /admin/biography:
    get:
      operationId: getAdminBiography
      summary: List all biography entries (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminBiographyEntry'
  /admin/biography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAdminBiographyById
      summary: Get single biography entry (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBiographyEntry'
        '404':
          description: Not found
  /admin/blog/categories:
    get:
      operationId: getAdminBlogCategories
      summary: List blog categories (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminBlogCategory'
    post:
      operationId: createAdminBlogCategory
      summary: Create blog category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBlogCategoryInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBlogCategory'
        '409':
          description: Conflict (duplicate name or slug)
...
  /admin/blog:
    get:
      operationId: getAdminBlog
      summary: List all blog posts including unpublished (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBlogPostList'
    post:
      operationId: createAdminBlogPost
      summary: Create blog post (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBlogPostInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBlogPost'
  /admin/blog/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAdminBlogPost
      summary: Get single blog post (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBlogPost'
        '404':
          description: Not found
    put:
      operationId: updateAdminBlogPost
      summary: Update blog post (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBlogPostInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBlogPost'
        '404':
          description: Not found
    delete:
      operationId: deleteAdminBlogPost
      summary: Delete blog post (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'